requirements.txt
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
decorator==5.1.0
dialogflow==1.0.0
Flask==2.0.2
future==0.18.2
google-api-core==1.31.4
google-auth==1.35.0
googleapis-common-protos==1.54.0
grpcio==1.42.0
idna==3.3
ipython==7.30.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.23
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
webexteamssdk==1.6
Werkzeug==2.0.2
